Move RandomOptimizer implementation to a new directory for optimizers
